Rename cli.py to pyproject_init.py
